GammaDist {stats}	R Documentation

The Gamma Distribution

Description

Density, distribution function, quantile function and random generation for the Gamma distribution with parameters shape and scale.

Usage

dgamma(x, shape, rate = 1, scale = 1/rate, log = FALSE)
pgamma(q, shape, rate = 1, scale = 1/rate, lower.tail = TRUE,
       log.p = FALSE)
qgamma(p, shape, rate = 1, scale = 1/rate, lower.tail = TRUE,
       log.p = FALSE)
rgamma(n, shape, rate = 1, scale = 1/rate)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

rate 	 an alternative way to specify the scale.

shape, scale 	 shape and scale parameters.

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].

Details

If scale is omitted, it assumes the default value of 1.

The Gamma distribution with parameters shape = a and scale = s has density

f(x)= 1/(s^a Gamma(a)) x^(a-1) e^-(x/s)

for x > 0, a > 0 and s > 0. The mean and variance are E(X) = a*s and Var(X) = a*s^2.

pgamma() uses a new algorithm (mainly by Morten Welinder) which should be uniformly better or equal to AS 239, see the references.

Value

dgamma gives the density, pgamma gives the distribution function qgamma gives the quantile function, and rgamma generates random deviates.

Note

The S parametrization is via shape and rate: S has no scale parameter.

The cumulative hazard H(t) = - log(1 - F(t)) is -pgamma(t, ..., lower = FALSE, log = TRUE).

pgamma is closely related to the incomplete gamma function. As defined by Abramowitz and Stegun 6.5.1

P(a,x) = 1/Gamma(a) integral_0^x t^(a-1) exp(-t) dt

P(a, x) is pgamma(x, a). Other authors (for example Karl Pearson in his 1922 tables) omit the normalizing factor, defining the incomplete gamma function as pgamma(x, a) * gamma(a).

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Shea, B. L. (1988) Algorithm AS 239, Chi-squared and Incomplete Gamma Integral, Applied Statistics (JRSS C) 37, 466–473.

Abramowitz, M. and Stegun, I. A. (1972) Handbook of Mathematical Functions. New York: Dover. Chapter 6: Gamma and Related Functions.

See Also

gamma for the Gamma function, dbeta for the Beta distribution and dchisq for the chi-squared distribution which is a special case of the Gamma distribution.

Examples

-log(dgamma(1:4, shape=1))
p <- (1:9)/10
pgamma(qgamma(p,shape=2), shape=2)
1 - 1/exp(qgamma(p, shape=1))

[Package stats version 2.1.0 Index]