IQR {stats}	R Documentation

The Interquartile Range

Description

computes interquartile range of the x values.

Usage


IQR(x, na.rm = FALSE, type = 7)

Arguments

x
a numeric vector.

na.rm
logical. Should missing values be removed?

type
an integer selecting one of the many quantile algorithms, see quantile.

Details

Note that this function computes the quartiles using the quantile function rather than following Tukey's recommendations, i.e., IQR(x) = quantile(x, 3/4) - quantile(x, 1/4).

For normally N(m,1) distributed X, the expected value of IQR(X) is 2*qnorm(3/4) = 1.3490, i.e., for a normal-consistent estimate of the standard deviation, use IQR(x) / 1.349.

References

Tukey, J. W. (1977). Exploratory Data Analysis. Reading: Addison-Wesley.

See Also

fivenum, mad which is more robust, range, quantile.

Examples


IQR(rivers)

[Package stats version 3.6.1 Index]