TDist {stats}	R Documentation

The Student t Distribution

Description

Density, distribution function, quantile function and random generation for the t distribution with df degrees of freedom (and optional noncentrality parameter ncp).

Usage

dt(x, df, ncp=0, log = FALSE)
pt(q, df, ncp=0, lower.tail = TRUE, log.p = FALSE)
qt(p, df,        lower.tail = TRUE, log.p = FALSE)
rt(n, df)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

df 	 degrees of freedom (> 0, maybe non-integer).

ncp 	 non-centrality parameter delta; currently for pt() and dt(), only for ncp <= 37.62.

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].

Details

The t distribution with df = n degrees of freedom has density

f(x) = Gamma((n+1)/2) / (sqrt(n pi) Gamma(n/2)) (1 + x^2/n)^-((n+1)/2)

for all real x. It has mean 0 (for n > 1) and variance n/(n-2) (for n > 2).

The general non-central t with parameters (df,Del) = (df, ncp) is defined as the distribution of T(df, Del) := (U + Del) / (Chi(df) / sqrt(df)) where U and Chi(df) are independent random variables, U ~ N(0,1), and Chi(df)^2 is chi-squared, see pchisq.

The most used applications are power calculations for t-tests:
Let T= (mX - m0) / (S/sqrt(n)) where mX is the mean and S the sample standard deviation (sd) of X_1,X_2,...,X_n which are i.i.d. N(mu,sigma^2). Then T is distributed as non-centrally t with df= n-1 degrees of freedom and non-centrality parameter ncp= (mu - m0) * sqrt(n)/sigma.

Value

dt gives the density, pt gives the distribution function, qt gives the quantile function, and rt generates random deviates.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole. (except non-central versions.)

Lenth, R. V. (1989). Algorithm AS 243 — Cumulative distribution function of the non-central t distribution, Appl. Statist. 38, 185–189.

See Also

df for the F distribution.

Examples

1 - pt(1:5, df = 1)
qt(.975, df = c(1:10,20,50,100,1000))

tt <- seq(0,10, len=21)
ncp <- seq(0,6, len=31)
ptn <- outer(tt,ncp, function(t,d) pt(t, df = 3, ncp=d))
image(tt,ncp,ptn, zlim=c(0,1),main=t.tit <- "Non-central t - Probabilities")
persp(tt,ncp,ptn, zlim=0:1, r=2, phi=20, theta=200, main=t.tit,
      xlab = "t", ylab = "noncentrality parameter", zlab = "Pr(T <= t)")

op <- par(yaxs="i")
plot(function(x) dt(x, df = 3, ncp = 2), -3, 11, ylim = c(0, 0.32),
     main="Non-central t - Density")
par(op)

[Package stats version 2.1.0 Index]