anova {stats}	R Documentation

Anova Tables

Description

Compute analysis of variance (or deviance) tables for one or more fitted model objects.

Usage


anova(object, ...)

Arguments

object
an object containing the results returned by a model fitting function (e.g., lm or glm).

...
additional objects of the same type.

Value

This (generic) function returns an object of class anova. These objects represent analysis-of-variance and analysis-of-deviance tables. When given a single argument it produces a table which tests whether the model terms are significant.

When given a sequence of objects, anova tests the models against one another in the order specified.

The print method for anova objects prints tables in a ‘pretty’ form.

Warning

The comparison between two or more models will only be valid if they are fitted to the same dataset. This may be a problem if there are missing values and R's default of na.action = na.omit is used.

References

Chambers, J. M. and Hastie, T. J. (1992) Statistical Models in S, Wadsworth & Brooks/Cole.

See Also

coefficients, effects, fitted.values, residuals, summary, drop1, add1.

[Package stats version 3.6.1 Index]