asOneSidedFormula {stats}	R Documentation

Convert to One-Sided Formula

Description

Names, expressions, numeric values, and character strings are converted to one-sided formulae. If object is a formula, it must be one-sided, in which case it is returned unaltered.

Usage


asOneSidedFormula(object)

Arguments

object
a one-sided formula, an expression, a numeric value, or a character string.

Value

a one-sided formula representing object

Author(s)

José Pinheiro and Douglas Bates

See Also

formula

Examples


asOneSidedFormula("age")
asOneSidedFormula(~ age)

[Package stats version 3.6.1 Index]