biplot {stats}	R Documentation

Biplot of Multivariate Data

Description

Plot a biplot on the current graphics device.

Usage


biplot(x, ...)

## Default S3 method:
biplot(x, y, var.axes = TRUE, col, cex = rep(par("cex"), 2),
       xlabs = NULL, ylabs = NULL, expand = 1,
       xlim  = NULL, ylim  = NULL, arrow.len = 0.1,
       main = NULL, sub = NULL, xlab = NULL, ylab = NULL, ...)

Arguments

x
The biplot, a fitted object. For biplot.default, the first set of points (a two-column matrix), usually associated with observations.

y
The second set of points (a two-column matrix), usually associated with variables.

var.axes
If TRUE the second set of points have arrows representing them as (unscaled) axes.

col
A vector of length 2 giving the colours for the first and second set of points respectively (and the corresponding axes). If a single colour is specified it will be used for both sets. If missing the default colour is looked for in the palette: if there it and the next colour as used, otherwise the first two colours of the palette are used.

cex
The character expansion factor used for labelling the points. The labels can be of different sizes for the two sets by supplying a vector of length two.

xlabs
A vector of character strings to label the first set of points: the default is to use the row dimname of x, or 1:n if the dimname is NULL.

ylabs
A vector of character strings to label the second set of points: the default is to use the row dimname of y, or 1:n if the dimname is NULL.

expand
An expansion factor to apply when plotting the second set of points relative to the first. This can be used to tweak the scaling of the two sets to a physically comparable scale.

arrow.len
The length of the arrow heads on the axes plotted in var.axes is true. The arrow head can be suppressed by arrow.len = 0.

xlim, ylim
Limits for the x and y axes in the units of the first set of variables.

main, sub, xlab, ylab, ...
graphical parameters.

Details

A biplot is plot which aims to represent both the observations and variables of a matrix of multivariate data on the same plot. There are many variations on biplots (see the references) and perhaps the most widely used one is implemented by biplot.princomp. The function biplot.default merely provides the underlying code to plot two sets of variables on the same figure.

Graphical parameters can also be given to biplot: the size of xlabs and ylabs is controlled by cex.

Side Effects

a plot is produced on the current graphics device.

References

K. R. Gabriel (1971). The biplot graphical display of matrices with application to principal component analysis. Biometrika, 58, 453–467. doi: 10.2307/2334381.

J.C. Gower and D. J. Hand (1996). Biplots. Chapman & Hall.

See Also

biplot.princomp, also for examples.

[Package stats version 3.6.1 Index]