kernapply {stats}	R Documentation

Apply Smoothing Kernel

Description

kernapply computes the convolution between an input sequence and a specific kernel.

Usage

kernapply(x, k, circular = FALSE, ...) 
kernapply(k1, k2)

Arguments

k, k1, k2 	 smoothing "tskernel" objects.

x 	 an input vector, matrix, or time series to be smoothed.

circular 	 a logical indicating whether the input sequence to be smoothed is treated as circular, i.e., periodic.

... 	 arguments passed to or from other methods.

Value

A smoothed version of the input sequence.

Author(s)

A. Trapletti

See Also

kernel, convolve, filter, spectrum

Examples

## see 'kernel' for examples

[Package stats version 2.1.0 Index]