lowess {stats}	R Documentation

Scatter Plot Smoothing

Description

This function performs the computations for the LOWESS smoother (see the reference below). lowess returns a list containing components x and y which give the coordinates of the smooth. The smooth should be added to a plot of the original points with the function lines.

Usage

lowess(x, y = NULL, f = 2/3, iter = 3,
       delta = 0.01 * diff(range(xy$x[o])))

Arguments

x, y 	 vectors giving the coordinates of the points in the scatter plot. Alternatively a single plotting structure can be specified.

f 	 the smoother span. This gives the proportion of points in the plot which influence the smooth at each value. Larger values give more smoothness.

iter 	 the number of robustifying iterations which should be performed. Using smaller values of iter will make lowess run faster.

delta 	 values of x which lie within delta of each other are replaced by a single value in the output from lowess. Defaults to 1/100th of the range of x.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Cleveland, W. S. (1979) Robust locally weighted regression and smoothing scatterplots. J. Amer. Statist. Assoc. 74, 829–836.

Cleveland, W. S. (1981) LOWESS: A program for smoothing scatterplots by robust locally weighted regression. The American Statistician, 35, 54.

See Also

loess), a newer formula based version of lowess (with different defaults!).

Examples

plot(cars, main = "lowess(cars)")
lines(lowess(cars), col = 2)
lines(lowess(cars, f=.2), col = 3)
legend(5, 120, c(paste("f = ", c("2/3", ".2"))), lty = 1, col = 2:3)

[Package stats version 2.1.0 Index]