make.link {stats}	R Documentation

Create a Link for GLM families

Description

This function is used with the family functions in glm(). Given a link, it returns a link function, an inverse link function, the derivative dmu/deta and a function for domain checking.

Usage

make.link(link)

Arguments

link 	 character or numeric; one of "logit", "probit", "cloglog", "identity", "log", "sqrt", "1/mu^2", "inverse", or number, say lambda resulting in power link mu ^ lambda.

Value

A list with components

linkfun 	 Link function function(mu)

linkinv 	 Inverse link function function(eta)

mu.eta 	 Derivative function(eta) dmu/deta

valideta 	 function(eta){ TRUE if all of eta is in the domain of linkinv }.

See Also

glm, family.

Examples

str(make.link("logit"))

l2 <- make.link(2)
l2$linkfun(0:3)# 0 1 4 9
l2$mu.eta(eta= 1:2)#= 1/(2*sqrt(eta))

[Package stats version 2.1.0 Index]