manova {stats}	R Documentation

Multivariate Analysis of Variance

Description

A class for the multivariate analysis of variance.

Usage

manova(...)

Arguments

... 	 Arguments to be passed to aov.

Details

Class "manova" differs from class "aov" in selecting a different summary method. Function manova calls aov and then add class "manova" to the result object for each stratum.

Value

See aov and the comments in ‘Details’ here.

Note

manova does not support multistratum analysis of variance, so the formula should not include an Error term.

References

Krzanowski, W. J. (1988) Principles of Multivariate Analysis. A User's Perspective. Oxford.

Hand, D. J. and Taylor, C. C. (1987) Multivariate Analysis of Variance and Repeated Measures. Chapman and Hall.

See Also

aov, summary.manova, the latter containing examples.

[Package stats version 2.9.2 Index]