naprint {stats}	R Documentation

Adjust for Missing Values

Description

Use missing value information to report the effects of an na.action.

Usage


naprint(x, ...)

Arguments

x
An object produced by an na.action function.

...
further arguments passed to or from other methods.

Details

This is a generic function, and the exact information differs by method. naprint.omit reports the number of rows omitted: naprint.default reports an empty string.

Value

A character string providing information on missing values, for example the number.

[Package stats version 3.6.1 Index]