nls.control {stats}	R Documentation

Control the Iterations in nls

Description

Allow the user to set some characteristics of the nls nonlinear least squares algorithm.

Usage


nls.control(maxiter = 50, tol = 1e-05, minFactor = 1/1024,
            printEval = FALSE, warnOnly = FALSE)

Arguments

maxiter
A positive integer specifying the maximum number of iterations allowed.

tol
A positive numeric value specifying the tolerance level for the relative offset convergence criterion.

minFactor
A positive numeric value specifying the minimum step-size factor allowed on any step in the iteration. The increment is calculated with a Gauss-Newton algorithm and successively halved until the residual sum of squares has been decreased or until the step-size factor has been reduced below this limit.

printEval
a logical specifying whether the number of evaluations (steps in the gradient direction taken each iteration) is printed.

warnOnly
a logical specifying whether nls() should return instead of signalling an error in the case of termination before convergence. Termination before convergence happens upon completion of maxiter iterations, in the case of a singular gradient, and in the case that the step-size factor is reduced below minFactor.

Value

A list with exactly five components:

maxiter

tol

minFactor

printEval

warnOnly

with meanings as explained under ‘Arguments’.

Author(s)

Douglas Bates and Saikat DebRoy

References

Bates, D. M. and Watts, D. G. (1988), Nonlinear Regression Analysis and Its Applications, Wiley.

See Also

nls

Examples


nls.control(minFactor = 1/2048)

[Package stats version 3.6.1 Index]