pairwise.prop.test {stats}	R Documentation

Pairwise comparisons for proportions

Description

Calculate pairwise comparisons between pairs of proportions with correction for multiple testing

Usage


pairwise.prop.test(x, n, p.adjust.method = p.adjust.methods, ...)

Arguments

x
Vector of counts of successes or a matrix with 2 columns giving the counts of successes and failures, respectively.

n
Vector of counts of trials; ignored if x is a matrix.

p.adjust.method
Method for adjusting p values (see p.adjust). Can be abbreviated.

...
Additional arguments to pass to prop.test

Value

Object of class "pairwise.htest"

See Also

prop.test, p.adjust

Examples


smokers  <- c( 83, 90, 129, 70 )
patients <- c( 86, 93, 136, 82 )
pairwise.prop.test(smokers, patients)

[Package stats version 3.6.1 Index]