plot.acf {stats}	R Documentation

Plot Autocovariance and Autocorrelation Functions

Description

Plot method for objects of class "acf".

Usage


## S3 method for class 'acf'
plot(x, ci = 0.95, type = "h", xlab = "Lag", ylab = NULL,
     ylim = NULL, main = NULL,
     ci.col = "blue", ci.type = c("white", "ma"),
     max.mfrow = 6, ask = Npgs > 1 && dev.interactive(),
     mar = if(nser > 2) c(3,2,2,0.8) else par("mar"),
     oma = if(nser > 2) c(1,1.2,1,1) else par("oma"),
     mgp = if(nser > 2) c(1.5,0.6,0) else par("mgp"),
     xpd = par("xpd"),
     cex.main = if(nser > 2) 1 else par("cex.main"),
     verbose = getOption("verbose"),
     ...)

Arguments

x
an object of class "acf".

ci
coverage probability for confidence interval. Plotting of the confidence interval is suppressed if ci is zero or negative.

type
the type of plot to be drawn, default to histogram like vertical lines.

xlab
the x label of the plot.

ylab
the y label of the plot.

ylim
numeric of length 2 giving the y limits for the plot.

main
overall title for the plot.

ci.col
colour to plot the confidence interval lines.

ci.type
should the confidence limits assume a white noise input or for lag k an MA(k-1) input? Can be abbreviated.

max.mfrow
positive integer; for multivariate x indicating how many rows and columns of plots should be put on one page, using par(mfrow = c(m,m)).

ask
logical; if TRUE, the user is asked before a new page is started.

mar, oma, mgp, xpd, cex.main
graphics parameters as in par(*), by default adjusted to use smaller than default margins for multivariate x only.

verbose
logical. Should R report extra information on progress?

...
graphics parameters to be passed to the plotting routines.

Note

The confidence interval plotted in plot.acf is based on an uncorrelated series and should be treated with appropriate caution. Using ci.type = "ma" may be less potentially misleading.

See Also

acf which calls plot.acf by default.

Examples


require(graphics)


z4  <- ts(matrix(rnorm(400), 100, 4), start = c(1961, 1), frequency = 12)
z7  <- ts(matrix(rnorm(700), 100, 7), start = c(1961, 1), frequency = 12)
acf(z4)
acf(z7, max.mfrow = 7)   # squeeze onto 1 page
acf(z7) # multi-page

[Package stats version 3.6.1 Index]