power {stats}	R Documentation

Create a Power Link Object

Description

Creates a link object based on the link function η = μ ^ λ.

Usage


power(lambda = 1)

Arguments

lambda
a real number.

Details

If lambda is non-positive, it is taken as zero, and the log link is obtained. The default lambda = 1 gives the identity link.

Value

A list with components linkfun, linkinv, mu.eta, and valideta. See make.link for information on their meaning.

References

Chambers, J. M. and Hastie, T. J. (1992) Statistical Models in S. Wadsworth & Brooks/Cole.

See Also

make.link, family

To raise a number to a power, see Arithmetic.

To calculate the power of a test, see various functions in the stats package, e.g., power.t.test.

Examples


power()
quasi(link = power(1/3))[c("linkfun", "linkinv")]

[Package stats version 3.6.1 Index]