print.ts {stats}	R Documentation

Printing and Formatting of Time-Series Objects

Description

Notably for calendar related time series objects, format and print methods showing years, months and or quarters respectively.

Usage


## S3 method for class 'ts'
print(x, calendar, ...)
.preformat.ts(x, calendar, ...)

Arguments

x
a time series object.

calendar
enable/disable the display of information about month names, quarter names or year when printing. The default is TRUE for a frequency of 4 or 12, FALSE otherwise.

...
additional arguments to print (or format methods).

Details

The print method for "ts" objects prints a header (basically of tsp(x)), if calendar is false, and then prints the result of .preformat.ts(x, *), which is typically a matrix with rownames built from the calendar times where applicable.

See Also

print, ts.

Examples


print(ts(1:10, frequency = 7, start = c(12, 2)), calendar = TRUE)

print(sunsp.1 <- window(sunspot.month, end=c(1756, 12)))
m <- .preformat.ts(sunsp.1) # a character matrix

[Package stats version 3.6.1 Index]