profiler {stats}	R Documentation

Constructor for Profiler Objects for Nonlinear Models

Description

Create a profiler object for the model object fitted.

Usage

profiler(fitted, ...)

Arguments

fitted 	 the original fitted model object.

... 	 Additional parameters. See documentation on individual methods.

Value

An object of class "profiler" which is a list with function elements

getFittedPars() 	 the parameters in fitted

setDefault(varying, params) 	 this is used for changing the default settings for profiling. In absence of both parameters, the default is set to the original fitted parameters with all parameters varying. The arguments are
varying: a logical, integer or character vector giving parameters to be varied. params: the default value at which profiling is to take place.

getProfile(varying, params) 	 this can be used in conjunction with setDefault without any arguments. Alternatively, the parameters to be varied and the values for fixed parameters can be specified using the arguments. The arguments are
varying: a logical vector giving parameters to be varied. This can be omitted if params is a named list or numeric vector.
params: values for parameters to be held fixed.
It returns a list with elements
parameters: the parameter values for the profiled optimum.
fstat: a profile statistics. See individual methods for details.
varying: a logical vector indicating parameters which were varied.

Author(s)

Douglas M. Bates and Saikat DebRoy

See Also

profiler.nls, profile

Examples

# see documentation on individual methods

[Package stats version 2.1.0 Index]