screeplot {stats}	R Documentation

Screeplot of PCA Results

Description

screeplot plots the variances against the number of the principal component. This is also the plot method for class "princomp".

Usage

screeplot(x, npcs = min(10, length(x$sdev)),
          type = c("barplot", "lines"),
          main = deparse(substitute(x)), ...)

Arguments

x 	 an object of class "princomp", as from princomp().

npcs 	 the number of principal components to be plotted.

type 	 the type of plot.

main, ... 	 graphics parameters.

References

Mardia, K. V., J. T. Kent and J. M. Bibby (1979). Multivariate Analysis, London: Academic Press.

Venables, W. N. and B. D. Ripley (2002). Modern Applied Statistics with S, Springer-Verlag.

See Also

princomp.

Examples

## The variances of the variables in the
## USArrests data vary by orders of magnitude, so scaling is appropriate
(pc.cr <- princomp(USArrests, cor = TRUE))  # inappropriate
screeplot(pc.cr)

fit <- princomp(covmat=Harman74.cor)
screeplot(fit)
screeplot(fit, npcs=24, type="lines")

[Package stats version 2.1.0 Index]