sd {stats}	R Documentation

Standard Deviation

Description

This function computes the standard deviation of the values in x. If na.rm is TRUE then missing values are removed before computation proceeds. If x is a matrix or a data frame, a vector of the standard deviation of the columns is returned.

Usage

sd(x, na.rm = FALSE)

Arguments

x 	 a numeric vector, matrix or data frame.

na.rm 	 logical. Should missing values be removed?

See Also

var for its square, and mad, the most robust alternative.

Examples

sd(1:2) ^ 2

[Package stats version 2.1.0 Index]