spec.taper {stats}	R Documentation

Taper a Time Series by a Cosine Bell

Description

Apply a cosine-bell taper to a time series.

Usage


spec.taper(x, p = 0.1)

Arguments

x
A univariate or multivariate time series

p
The proportion to be tapered at each end of the series, either a scalar (giving the proportion for all series) or a vector of the length of the number of series (giving the proportion for each series).

Details

The cosine-bell taper is applied to the first and last p[i] observations of time series x[, i].

Value

A new time series object.

See Also

spec.pgram, cpgram

[Package stats version 3.6.1 Index]