stat.anova {stats}	R Documentation

GLM Anova Statistics

Description

This is a utility function, used in lm and glm methods for anova(..., test != NULL) and should not be used by the average user.

Usage


stat.anova(table, test = c("Rao","LRT", "Chisq", "F", "Cp"),
           scale, df.scale, n)

Arguments

table
numeric matrix as results from anova.glm(..., test = NULL).

test
a character string, partially matching one of "Rao", "LRT", "Chisq", "F" or "Cp".

scale
a residual mean square or other scale estimate to be used as the denominator in an F test.

df.scale
degrees of freedom corresponding to scale.

n
number of observations.

Value

A matrix which is the original table, augmented by a column of test statistics, depending on the test argument.

References

Hastie, T. J. and Pregibon, D. (1992) Generalized linear models. Chapter 6 of Statistical Models in S eds J. M. Chambers and T. J. Hastie, Wadsworth & Brooks/Cole.

See Also

anova.lm, anova.glm.

Examples


##-- Continued from '?glm':

print(ag <- anova(glm.D93))
stat.anova(ag$table, test = "Cp",
           scale = sum(resid(glm.D93, "pearson")^2)/4,
           df.scale = 4, n = 9)

[Package stats version 3.6.1 Index]