stats-defunct {stats}	R Documentation

Defunct Functions in Package stats

Description

The functions or variables listed here are no longer part of R as they are not needed (any more).

Usage


## Defunct in R 1.x
dnchisq(.)
pnchisq(.)
qnchisq(.)
rnchisq(.)
print.anova.glm(.)
print.anova.lm(.)
reshapeWide(x, i, j, val, jnames = levels(j))
reshapeLong(x, jvars,  ilev = row.names(x),
            jlev = names(x)[jvars], iname = "reshape.i",
            jname = "reshape.j", vname = "reshape.v")
arima0.diag(...)
plot.mts(x, plot.type = c("multiple", "single"), panel = lines,
         log = "", col = par("col"),  bg = NA, pch = par("pch"),
         cex = par("cex"), lty = par("lty"), lwd = par("lwd"),
         ann = par("ann"),  xlab = "Time", type = "l", main = NULL,
         oma = c(6, 0, 5, 0), ...)
print.coefmat(x, digits = max(3, getOption("digits") - 2),
              signif.stars = getOption("show.signif.stars"),
              dig.tst = max(1, min(5, digits - 1)),
              cs.ind, tst.ind, zap.ind = integer(0),
              P.values = NULL, has.Pvalue,
              eps.Pvalue = .Machine$double.eps,
              na.print = "", ...)
anovalist.lm(object, ..., test = NULL)
lm.fit.null(x, y, method = "qr", tol = 1e-07, ...)
lm.wfit.null(x, y, w, method = "qr", tol = 1e-07, ...)
glm.fit.null(x, y, weights, start = NULL,
             etastart = NULL, mustart = NULL, offset,
             family = gaussian(), control = glm.control(),
             intercept = FALSE)

# Defunct in 2.4.0
mauchley.test(object, Sigma = diag(nrow = p),
   T = Thin.row(proj(M) - proj(X)), M = diag(nrow = p), X = ~0,
   idata = data.frame(index = seq(length = p)), ...)

# Defunct in 2.10.0
clearNames(object)

Details

The *chisq() functions now take an optional non-centrality argument, so the *nchisq() functions are no longer needed.

reshape*, which were experimental, are replaced by reshape. This has a different syntax and allows multiple time-varying variables.

arima0.diag has been replaced by tsdiag.arima0.

plot.mts has been removed, as plot.ts now has the same functionality.

print.coefmat was an older name for printCoefmat with a different default for na.print.

anovalist.lm was replaced by anova.lmlist in R 1.2.0.

lm.fit.null and lm.wfit.null are superseded by lm.fit and lm.wfit which handle null models now. Similarly, glm.fit.null is superseded by glm.fit.

mauchley.test was a misspelling of Mauchly's name, corrected by the introduction of mauchly.test.

clearNames had been introduced at about the same time as unname, but is less general and has been used rarely.

See Also

Defunct

[Package stats version 3.6.1 Index]