stats-deprecated {stats}	R Documentation

Deprecated Functions in Package stats

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as the next release.

Usage


plclust(tree, hang = 0.1, unit = FALSE, level = FALSE, hmin = 0,
        square = TRUE, labels = NULL, plot. = TRUE,
        axes = TRUE, frame.plot = FALSE, ann = TRUE,
        main = "", sub = NULL, xlab = NULL, ylab = "Height")

Arguments

tree
an object of the type produced by hclust.

hang
The fraction of the plot height by which labels should hang below the rest of the plot. A negative value will cause the labels to hang down from 0.

unit
logical. If true, the splits are plotted at equally-spaced heights rather than at the height in the object.

labels
A character vector of labels for the leaves of the tree. By default the row names or row numbers of the original data are used. If labels = FALSE no labels at all are plotted.

axes, frame.plot, ann
logical flags as in plot.default.

main, sub, xlab, ylab
character strings for title. sub and xlab have a non-NULL default when there's a tree$call.

...
Further graphical arguments. E.g., cex controls the size of the labels (if plotted) in the same way as text.

hmin
numeric. All heights less than hmin are regarded as being hmin: this can be used to suppress detail at the bottom of the tree.

level, square, plot.
unimplemented arguments for S-PLUS compatibility.

Details

plcust is a deprecated wrapper for the plot method for hclust, provided long ago for S-PLUS compatibility.

See Also

Deprecated

[Package stats version 3.6.1 Index]