stlmethods {stats}	R Documentation

Methods for STL Objects

Description

Methods for objects of class stl, typically the result of stl. The plot method does a multiple figure plot with some flexibility.

There are also (non-visible) print and summary methods.

Usage


## S3 method for class 'stl'
plot(x, labels = colnames(X),
     set.pars = list(mar = c(0, 6, 0, 6), oma = c(6, 0, 4, 0),
                     tck = -0.01, mfrow = c(nplot, 1)),
     main = NULL, range.bars = TRUE, ...,
     col.range = "light gray")

Arguments

x
stl object.

labels
character of length 4 giving the names of the component time-series.

set.pars
settings for par(.) when setting up the plot.

main
plot main title.

range.bars
logical indicating if each plot should have a bar at its right side which are of equal heights in user coordinates.

...
further arguments passed to or from other methods.

col.range
colour to be used for the range bars, if plotted. Note this appears after ... and so cannot be abbreviated.

See Also

plot.ts and stl, particularly for examples.

[Package stats version 3.6.1 Index]