ts-defunct {stats}	R Documentation

Defunct functions in Package ts

Description

The functions or variables listed here are no longer part of R as they are not needed (any more).

Usage

arima0.diag(...)

Details

arima0.diag has been replaced by tsdiag.arima0.

[Package stats version 2.9.2 Index]