weighted.residuals {stats}	R Documentation

Compute Weighted Residuals

Description

Computed weighted residuals from a linear model fit.

Usage


weighted.residuals(obj, drop0 = TRUE)

Arguments

obj
R object, typically of class lm or glm.

drop0
logical. If TRUE, drop all cases with weights == 0.

Details

Weighted residuals are based on the deviance residuals, which for a lm fit are the raw residuals Ri multiplied by wi^0.5, where wi are the weights as specified in lm's call.

Dropping cases with weights zero is compatible with influence and related functions.

Value

Numeric vector of length n', where n' is the number of of non-0 weights (drop0 = TRUE) or the number of observations, otherwise.

See Also

residuals, lm.influence, etc.

Examples


## following on from example(lm)

all.equal(weighted.residuals(lm.D9),
          residuals(lm.D9))
x <- 1:10
w <- 0:9
y <- rnorm(x)
weighted.residuals(lmxy <- lm(y ~ x, weights = w))
weighted.residuals(lmxy, drop0 = FALSE)

[Package stats version 3.6.1 Index]