Lognormal {stats}	R Documentation

The Log Normal Distribution

Description

Density, distribution function, quantile function and random generation for the log normal distribution whose logarithm has mean equal to meanlog and standard deviation equal to sdlog.

Usage

dlnorm(x, meanlog = 0, sdlog = 1, log = FALSE)
plnorm(q, meanlog = 0, sdlog = 1, lower.tail = TRUE, log.p = FALSE)
qlnorm(p, meanlog = 0, sdlog = 1, lower.tail = TRUE, log.p = FALSE)
rlnorm(n, meanlog = 0, sdlog = 1)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

meanlog, sdlog 	 mean and standard deviation of the distribution on the log scale with default values of 0 and 1 respectively.

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].

Details

The log normal distribution has density

f(x) = 1/(sqrt(2 pi) sigma x) e^-((log x - mu)^2 / (2 sigma^2))

where μ and σ are the mean and standard deviation of the logarithm. The mean is E(X) = exp(μ + 1/2 σ^2), the median is med(X) = exp(μ), and the variance Var(X) = exp(2*mu + sigma^2)*(exp(sigma^2) - 1) and hence the coefficient of variation is sqrt(exp(sigma^2) - 1) which is approximately σ when that is small (e.g., σ < 1/2).

Value

dlnorm gives the density, plnorm gives the distribution function, qlnorm gives the quantile function, and rlnorm generates random deviates.

Note

The cumulative hazard H(t) = - log(1 - F(t)) is -plnorm(t, r, lower = FALSE, log = TRUE).

Source

dlnorm is calculated from the definition (in ‘Details’). [pqr]lnorm are based on the relationship to the normal.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 1, chapter 14. Wiley, New York.

See Also

dnorm for the normal distribution.

Examples

dlnorm(1) == dnorm(0)

[Package stats version 2.9.2 Index]