deviance {stats}	R Documentation

Model Deviance

Description

Returns the deviance of a fitted model object.

Usage

deviance(object, ...)

Arguments

object 	 an object for which the deviance is desired.

... 	 additional optional argument.

Details

This is a generic function which can be used to extract deviances for fitted models. Consult the individual modeling functions for details on how to use this function.

Value

The value of the deviance extracted from the object object.

References

Chambers, J. M. and Hastie, T. J. (1992) Statistical Models in S. Wadsworth & Brooks/Cole.

See Also

df.residual, extractAIC, glm, lm.

[Package stats version 2.9.2 Index]