diffinv {stats}	R Documentation

Discrete Integration: Inverse of Differencing

Description

Computes the inverse function of the lagged differences function diff.

Usage

diffinv(x, ...)

## Default S3 method:
diffinv(x, lag = 1, differences = 1, xi, ...)
## S3 method for class 'ts':
diffinv(x, lag = 1, differences = 1, xi, ...)

Arguments

x 	 a numeric vector, matrix, or time series.

lag 	 a scalar lag parameter.

differences 	 an integer representing the order of the difference.

xi 	 a numeric vector, matrix, or time series containing the initial values for the integrals. If missing, zeros are used.

... 	 arguments passed to or from other methods.

Details

diffinv is a generic function with methods for class "ts" and default for vectors and matrices.

Missing values are not handled.

Value

A numeric vector, matrix, or time series (the latter for the "ts" method) representing the discrete integral of x.

Author(s)

A. Trapletti

See Also

diff

Examples

s <- 1:10
d <- diff(s)
diffinv(d, xi = 1)

[Package stats version 2.9.2 Index]