glm.control {stats}	R Documentation

Auxiliary for Controlling GLM Fitting

Description

Auxiliary function as user interface for glm fitting. Typically only used when calling glm or glm.fit.

Usage

glm.control(epsilon = 1e-8, maxit = 25, trace = FALSE)

Arguments

epsilon 	 positive convergence tolerance ε; the iterations converge when |dev - devold|/(|dev| + 0.1) < ε.

maxit 	 integer giving the maximal number of IWLS iterations.

trace 	 logical indicating if output should be produced for each iteration.

Details

If epsilon is small, it is also used as the tolerance for the least squares solution.

When trace is true, calls to cat produce the output for each IWLS iteration. Hence, options(digits = *) can be used to increase the precision, see the example.

Value

A list with the arguments as components.

References

Hastie, T. J. and Pregibon, D. (1992) Generalized linear models. Chapter 6 of Statistical Models in S eds J. M. Chambers and T. J. Hastie, Wadsworth & Brooks/Cole.

See Also

glm.fit, the fitting procedure used by glm.

Examples

### A variation on  example(glm) :

## Annette Dobson's example ...
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
oo <- options(digits = 12) # to see more when tracing :
glm.D93X <- glm(counts ~ outcome + treatment, family=poisson(),
                trace = TRUE, epsilon = 1e-14)
options(oo)
coef(glm.D93X) # the last two are closer to 0 than in ?glm's  glm.D93

[Package stats version 2.9.2 Index]