is.empty.model {stats}	R Documentation

Check if a Model is Empty

Description

R model notation allows models with no intercept and no predictors. These require special handling internally. is.empty.model() checks whether an object describes an empty model.

Usage

is.empty.model(x)

Arguments

x 	 A terms object or an object with a terms method.

Value

TRUE if the model is empty

See Also

lm,glm

Examples

y <- rnorm(20)
is.empty.model(y ~ 0)
is.empty.model(y ~ -1)
is.empty.model(lm(y ~ 0))

[Package stats version 2.9.2 Index]