kernapply {stats}	R Documentation

Apply Smoothing Kernel

Description

kernapply computes the convolution between an input sequence and a specific kernel.

Usage


kernapply(x, ...)

## Default S3 method:
kernapply(x, k, circular = FALSE, ...)
## S3 method for class 'ts'
kernapply(x, k, circular = FALSE, ...)
## S3 method for class 'vector'
kernapply(x, k, circular = FALSE, ...)

## S3 method for class 'tskernel'
kernapply(x, k, ...)

Arguments

x
an input vector, matrix, time series or kernel to be smoothed.

k
smoothing "tskernel" object.

circular
a logical indicating whether the input sequence to be smoothed is treated as circular, i.e., periodic.

...
arguments passed to or from other methods.

Value

A smoothed version of the input sequence.

Note

This uses fft to perform the convolution, so is fastest when NROW(x) is a power of 2 or some other highly composite integer.

Author(s)

A. Trapletti

See Also

kernel, convolve, filter, spectrum

Examples


## see 'kernel' for examples

[Package stats version 3.6.1 Index]