listof {stats}	R Documentation

A Class for Lists of (Parts of) Model Fits

Description

Class "listof" is used by aov and the "lm" method of alias for lists of model fits or parts thereof. It is simply a list with an assigned class to control the way methods, especially printing, act on it.

It has a coef method in this package (which returns an object of this class), and [ and print methods in package base.

[Package stats version 4.1.2 Index]