loadings {stats}	R Documentation

Print Loadings in Factor Analysis

Description

Extract or print loadings in factor analysis (or principal components analysis).

Usage

loadings(x)

## S3 method for class 'loadings':
print(x, digits = 3, cutoff = 0.1, sort = FALSE, ...)

## S3 method for class 'factanal':
print(x, digits = 3, ...)

Arguments

x 	 an object of class "factanal" or "princomp" or the loadings component of such an object.

digits 	 number of decimal places to use in printing uniquenesses and loadings.

cutoff 	 loadings smaller than this (in absolute value) are suppressed.

sort 	 logical. If true, the variables are sorted by their importance on each factor. Each variable with any loading larger than 0.5 (in modulus) is assigned to the factor with the largest loading, and the variables are printed in the order of the factor they are assigned to, then those unassigned.

... 	 further arguments for other methods, such as cutoff and sort for print.factanal.

See Also

factanal, princomp

[Package stats version 2.9.2 Index]