loess.control {stats}	R Documentation

Set Parameters for Loess

Description

Set control parameters for loess fits.

Usage

loess.control(surface = c("interpolate", "direct"),
              statistics = c("approximate", "exact"),
              trace.hat = c("exact", "approximate"),
              cell = 0.2, iterations = 4, ...)

Arguments

surface 	 should be fitted surface be computed exactly or via interpolation from a kd tree?

statistics 	 should the statistics be computed exactly or approximately? Exact computation can be very slow.

trace.hat 	 should the trace of the smoother matrix be computed exactly or approximately? It is recommended to use the approximation for more than about 1000 data points.

cell 	 if interpolation is used this controls the accuracy of the approximation via the maximum number of points in a cell in the kd tree. Cells with more than floor(n*span*cell) points are subdivided.

iterations 	 the number of iterations used in robust fitting.

... 	 further arguments which are ignored.

Value

A list with components

surface

statistics

trace.hat

cell

iterations

with meanings as explained under ‘Arguments’.

See Also

loess

[Package stats version 2.9.2 Index]