loess.control {stats}	R Documentation

Set Parameters for Loess

Description

Set control parameters for loess fits.

Usage


loess.control(surface = c("interpolate", "direct"),
              statistics = c("approximate", "exact", "none"),
              trace.hat = c("exact", "approximate"),
              cell = 0.2, iterations = 4, iterTrace = FALSE, ...)

Arguments

surface
should the fitted surface be computed exactly ("direct") or via interpolation from a kd tree? Can be abbreviated.

statistics
should the statistics be computed exactly, approximately or not at all? Exact computation can be very slow. Can be abbreviated.

trace.hat
Only for the (default) case (surface = "interpolate", statistics = "approximate"): should the trace of the smoother matrix be computed exactly or approximately? It is recommended to use the approximation for more than about 1000 data points. Can be abbreviated.

cell
if interpolation is used this controls the accuracy of the approximation via the maximum number of points in a cell in the kd tree. Cells with more than floor(n*span*cell) points are subdivided.

iterations
the number of iterations used in robust fitting, i.e. only if family is "symmetric".

iterTrace
logical (or integer) determining if tracing information during the robust iterations (iterations>= 2) is produced.

...
further arguments which are ignored.

Value

A list with components

surface

statistics

trace.hat

cell

iterations

iterTrace

with meanings as explained under ‘Arguments’.

See Also

loess

[Package stats version 3.6.1 Index]