ls.print {stats}	R Documentation

Print ‘lsfit’ Regression Results

Description

Computes basic statistics, including standard errors, t- and p-values for the regression coefficients and prints them if print.it is TRUE.

Usage

ls.print(ls.out, digits = 4, print.it = TRUE)

Arguments

ls.out 	 Typically the result of lsfit()

digits 	 The number of significant digits used for printing

print.it 	 a logical indicating whether the result should also be printed

Value

A list with the components

summary 	 The ANOVA table of the regression

coef.table 	 matrix with regression coefficients, standard errors, t- and p-values

Note

Usually you would use summary(lm(...)) and anova(lm(...)) to obtain similar output.

See Also

ls.diag, lsfit, also for examples; lm, lm.influence which usually are preferable.

[Package stats version 2.9.2 Index]