median {stats}	R Documentation

Median Value

Description

Compute the sample median.

Usage

median(x, na.rm = FALSE)

Arguments

x 	 an object for which a method has been defined, or a numeric vector containing the values whose median is to be computed.

na.rm 	 a logical value indicating whether NA values should be stripped before the computation proceeds.

Details

This is a generic function for which methods can be written. However, the default method makes use of sort and mean from package base both of which are generic, and so the default method will work for most classes (e.g. "Date") for which a median is a reasonable concept.

Value

The default method returns a length-one object of the same type as x, except when x is integer of even length, when the result will be double.
If there are no values or if na.rm = FALSE and there are NA values the result is NA of the same type as x (or more generally the result of x[FALSE][NA]).

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

quantile for general quantiles.

Examples

median(1:4)# = 2.5 [even number]
median(c(1:3,100,1000))# = 3 [odd, robust]

[Package stats version 2.9.2 Index]