na.action {stats}	R Documentation

NA Action

Description

Extract information on the NA action used to create an object.

Usage

na.action(object, ...)

Arguments

object 	 any object whose NA action is given.

... 	 further arguments special methods could require.

Details

na.action is a generic function, and na.action.default its default method. The latter extracts the "na.action" component of a list if present, otherwise the "na.action" attribute.

When model.frame is called, it records any information on NA handling in a "na.action" attribute. Most model-fitting functions return this as a component of their result.

Value

Information from the action which was applied to object if NAs were handled specially, or NULL.

References

Chambers, J. M. and Hastie, T. J. (1992) Statistical Models in S. Wadsworth & Brooks/Cole.

See Also

options("na.action"), na.omit, na.fail, also for na.exclude, na.pass.

Examples

na.action(na.omit(c(1, NA)))

[Package stats version 2.9.2 Index]