naresid {stats}	R Documentation

Adjust for Missing Values

Description

Use missing value information to adjust residuals and predictions.

Usage


naresid(omit, x, ...)
napredict(omit, x, ...)

Arguments

omit
an object produced by an na.action function, typically the "na.action" attribute of the result of na.omit or na.exclude.

x
a vector, data frame, or matrix to be adjusted based upon the missing value information.

...
further arguments passed to or from other methods.

Details

These are utility functions used to allow predict, fitted and residuals methods for modelling functions to compensate for the removal of NAs in the fitting process. They are used by the default, "lm", "glm" and "nls" methods, and by further methods in packages MASS, rpart and survival. Also used for the scores returned by factanal, prcomp and princomp.

The default methods do nothing. The default method for the na.exclude action is to pad the object with NAs in the correct positions to have the same number of rows as the original data frame.

Currently naresid and napredict are identical, but future methods need not be. naresid is used for residuals, and napredict for fitted values, predictions and weights.

Value

These return a similar object to x.

Note

In the early 2000s, packages rpart and survival5 contained versions of these functions that had an na.omit action equivalent to that now used for na.exclude.

[Package stats version 3.6.1 Index]