nextn {stats}	R Documentation

Highly Composite Numbers

Description

nextn returns the smallest integer, greater than or equal to n, which can be obtained as a product of powers of the values contained in factors. nextn is intended to be used to find a suitable length to zero-pad the argument of fft to so that the transform is computed quickly. The default value for factors ensures this.

Usage

nextn(n, factors = c(2,3,5))

Arguments

n 	 an integer.

factors 	 a vector of positive integer factors.

See Also

convolve, fft.

Examples

nextn(1001) # 1024
table(sapply(599:630, nextn))

[Package stats version 2.9.2 Index]