pairwise.table {stats}	R Documentation

Tabulate p values for pairwise comparisons

Description

Creates table of p values for pairwise comparisons with corrections for multiple testing.

Usage


pairwise.table(compare.levels, level.names, p.adjust.method)

Arguments

compare.levels
a function to compute (raw) p value given indices i and j.

level.names
names of the group levels

p.adjust.method
a character string specifying the method for multiple testing adjustment; almost always one of p.adjust.methods. Can be abbreviated.

Details

Functions that do multiple group comparisons create separate compare.levels functions (assumed to be symmetrical in i and j) and passes them to this function.

Value

Table of p values in lower triangular form.

See Also

pairwise.t.test

[Package stats version 3.6.1 Index]