plot.HoltWinters {stats}	R Documentation

Plot function for HoltWinters objects

Description

Produces a chart of the original time series along with the fitted values. Optionally, predicted values (and their confidence bounds) can also be plotted.

Usage

## S3 method for class 'HoltWinters':
plot(x, predicted.values = NA, intervals = TRUE,
        separator = TRUE, col = 1, col.predicted = 2,
        col.intervals = 4, col.separator = 1, lty = 1,
        lty.predicted = 1, lty.intervals = 1, lty.separator = 3,
        ylab = "Observed / Fitted",
        main = "Holt-Winters filtering",
        ylim = NULL, ...)

Arguments

x 	 Object of class "HoltWinters"

predicted.values 	 Predicted values as returned by predict.HoltWinters

intervals 	 If TRUE, the prediction intervals are plotted (default).

separator 	 If TRUE, a separating line between fitted and predicted values is plotted (default).

col, lty 	 Color/line type of original data (default: black solid).

col.predicted, lty.predicted 	 Color/line type of fitted and predicted values (default: red solid).

col.intervals, lty.intervals 	 Color/line type of prediction intervals (default: blue solid).

col.separator, lty.separator 	 Color/line type of observed/predicted values separator (default: black dashed).

ylab 	 Label of the y-axis.

main 	 Main title.

ylim 	 Limits of the y-axis. If NULL, the range is chosen such that the plot contains the original series, the fitted values, and the predicted values if any.

... 	 Other graphics parameters.

Author(s)

David Meyer David.Meyer@wu-wien.ac.at

References

C. C. Holt (1957) Forecasting trends and seasonals by exponentially weighted moving averages, ONR Research Memorandum, Carnegie Institute of Technology 52.

P. R. Winters (1960) Forecasting sales by exponentially weighted moving averages, Management Science 6, 324–342.

See Also

HoltWinters, predict.HoltWinters

[Package stats version 2.9.2 Index]