poisson.test {stats}	R Documentation

Exact Poisson tests

Description

Performs an exact test of a simple null hypothesis about the rate parameter in Poisson distribution, or for the ratio between two rate parameters.

Usage

poisson.test(x, T = 1, r = 1,
    alternative = c("two.sided", "less", "greater"),
    conf.level = 0.95)

Arguments

x 	 number of events. A vector of length one or two.

T 	 time base for event count. A vector of length one or two.

r 	 hypothesized rate or rate ratio

alternative 	 indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". You can specify just the initial letter.

conf.level 	 confidence level for the returned confidence interval.

Details

Confidence intervals are computed similarly to those of binom.test in the one-sample case, and using binom.test in the two sample case.

Value

A list with class "htest" containing the following components:

statistic 	 the number of events (in the first sample if there are two.)

parameter 	 the corresponding expected count

p.value 	 the p-value of the test.

conf.int 	 a confidence interval for the rate or rate ratio.

estimate 	 the estimated rate or rate ratio.

null.value 	 the rate or rate ratio under the null, r.

alternative 	 a character string describing the alternative hypothesis.

method 	 the character string "Exact Poisson test" or "Comparison of Poisson rates" as appropriate.

data.name 	 a character string giving the names of the data.

Note

The rate parameter in Poisson data is often given based on a “time on test” or similar quantity (person-years, population size, or expected number of cases from mortality tables). This is the role of the T argument.

The one-sample case is effectively the binomial test with a very large n. The two sample case is converted to a binomial test by conditioning on the total event count, and the rate ratio is directly related to the odds in that binomial distribution.

See Also

binom.test

Examples

### These are paraphrased from data sets in the ISwR package

## SMR, Welsh Nickel workers
poisson.test(137, 24.19893)

## eba1977, compare Fredericia to other three cities for ages 55-59
poisson.test(c(11,6+8+7),c(800, 1083+1050+878))

[Package stats version 2.9.2 Index]