power.prop.test {stats}	R Documentation

Power calculations two sample test for proportions

Description

Compute power of test, or determine parameters to obtain target power.

Usage

power.prop.test(n = NULL, p1 = NULL, p2 = NULL, sig.level = 0.05,
                power = NULL,
                alternative = c("two.sided", "one.sided"),
                strict = FALSE)

Arguments

n 	 Number of observations (per group)

p1 	 probability in one group

p2 	 probability in other group

sig.level 	 Significance level (Type I error probability)

power 	 Power of test (1 minus Type II error probability)

alternative 	 One- or two-sided test

strict 	 Use strict interpretation in two-sided case

Details

Exactly one of the parameters n, p1, p2, power, and sig.level must be passed as NULL, and that parameter is determined from the others. Notice that sig.level has a non-NULL default so NULL must be explicitly passed if you want it computed.

If strict = TRUE is used, the power will include the probability of rejection in the opposite direction of the true effect, in the two-sided case. Without this the power will be half the significance level if the true difference is zero.

Value

Object of class "power.htest", a list of the arguments (including the computed one) augmented with method and note elements.

Note

uniroot is used to solve power equation for unknowns, so you may see errors from it, notably about inability to bracket the root when invalid arguments are given. If one of them is computed p1 < p2 will hold, although this is not enforced when both are specified.

Author(s)

Peter Dalgaard. Based on previous work by Claus Ekstrøm

See Also

prop.test, uniroot

Examples

 power.prop.test(n = 50, p1 = .50, p2 = .75)
 power.prop.test(p1 = .50, p2 = .75, power = .90)
 power.prop.test(n = 50, p1 = .5, power = .90)

[Package stats version 2.9.2 Index]