preplot {stats}	R Documentation

Pre-computations for a Plotting Objeect

Description

Compute an object to be used for plots relating to the given model object.

Usage

preplot(object, ...)

Arguments

object 	 a fitted model object.

... 	 additional arguments for specific methods.

Details

Only the generic function is currently provided in base R, but some add-on packages have methods. Principally here for S compatibility.

Value

An object set up to make a plot that describes object.

[Package stats version 2.9.2 Index]