print.power.htest {stats}	R Documentation

Print method for power calculation object

Description

Print object of class "power.htest" in nice layout.

Usage

## S3 method for class 'power.htest':
print(x, ...)

Arguments

x 	 Object of class "power.htest".

... 	 further arguments to be passed to or from methods.

Details

A power.htest object is just a named list of numbers and character strings, supplemented with method and note elements. The method is displayed as a title, the note as a footnote, and the remaining elements are given in an aligned ‘name = value’ format.

Value

none

Author(s)

Peter Dalgaard

See Also

power.t.test, power.prop.test

[Package stats version 2.9.2 Index]