print.ts {stats}	R Documentation

Printing Time-Series Objects

Description

Print method for time series objects.

Usage

## S3 method for class 'ts':
print(x, calendar, ...)

Arguments

x 	 a time series object.

calendar 	 enable/disable the display of information about month names, quarter names or year when printing. The default is TRUE for a frequency of 4 or 12, FALSE otherwise.

... 	 additional arguments to print.

Details

This is the print methods for objects inheriting from class "ts".

See Also

print, ts.

Examples

print(ts(1:10, frequency = 7, start = c(12, 2)), calendar = TRUE)

[Package stats version 2.9.2 Index]