printCoefmat {stats}	R Documentation

Print Coefficient Matrices

Description

Utility function to be used in higher-level print methods, such as print.summary.lm, print.summary.glm and print.anova. The goal is to provide a flexible interface with smart defaults such that often, only x needs to be specified.

Usage

printCoefmat(x, digits=max(3, getOption("digits") - 2),
             signif.stars = getOption("show.signif.stars"),
             signif.legend = signif.stars,
             dig.tst = max(1, min(5, digits - 1)),
             cs.ind = 1L:k, tst.ind = k + 1L, zap.ind = integer(0),
             P.values = NULL,
             has.Pvalue = nc >= 4L &&
                          substr(colnames(x)[nc], 1L, 3L) == "Pr(",
             eps.Pvalue = .Machine$double.eps,
             na.print = "NA", ...)

Arguments

x 	 a numeric matrix like object, to be printed.

digits 	 minimum number of significant digits to be used for most numbers.

signif.stars 	 logical; if TRUE, P-values are additionally encoded visually as ‘significance stars’ in order to help scanning of long coefficient tables. It defaults to the show.signif.stars slot of options.

signif.legend 	 logical; if TRUE, a legend for the ‘significance stars’ is printed provided signif.stars=TRUE.

dig.tst 	 minimum number of significant digits for the test statistics, see tst.ind.

cs.ind 	 indices (integer) of column numbers which are (like) coefficients and standard errors to be formatted together.

tst.ind 	 indices (integer) of column numbers for test statistics.

zap.ind 	 indices (integer) of column numbers which should be formatted by zapsmall, i.e., by ‘zapping’ values close to 0.

P.values 	 logical or NULL; if TRUE, the last column of x is formatted by format.pval as P values. If P.values = NULL, the default, it is set to TRUE only if options("show.coef.Pvalue") is TRUE and x has at least 4 columns and the last column name of x starts with "Pr(".

has.Pvalue 	 logical; if TRUE, the last column of x contains P values; in that case, it is printed if and only if P.values (above) is true.

eps.Pvalue 	 number,..

na.print 	 a character string to code NA values in printed output.

... 	 further arguments for print.

Value

Invisibly returns its argument, x.

Author(s)

Martin Maechler

See Also

print.summary.lm, format.pval, format.

Examples

cmat <- cbind(rnorm(3, 10), sqrt(rchisq(3, 12)))
cmat <- cbind(cmat, cmat[,1]/cmat[,2])
cmat <- cbind(cmat, 2*pnorm(-cmat[,3]))
colnames(cmat) <- c("Estimate", "Std.Err", "Z value", "Pr(>z)")
printCoefmat(cmat[,1:3])
printCoefmat(cmat)
options(show.coef.Pvalues = FALSE)
printCoefmat(cmat, digits=2)
printCoefmat(cmat, digits=2, P.values = TRUE)
options(show.coef.Pvalues = TRUE)# revert

[Package stats version 2.9.2 Index]