profile {stats}	R Documentation

Generic Function for Profiling Models

Description

Investigates behavior of objective function near the solution represented by fitted.

See documentation on method functions for further details.

Usage

profile(fitted, ...)

Arguments

fitted 	 the original fitted model object.

... 	 additional parameters. See documentation on individual methods.

Value

A list with an element for each parameter being profiled. See the individual methods for further details.

See Also

profile.nls, profile.glm in package MASS, ...

For profiling R code, see Rprof.

[Package stats version 2.9.2 Index]