profile.nls {stats}	R Documentation

Method for Profiling nls Objects

Description

Investigates the profile log-likelihood function for a fitted model of class "nls".

Usage

## S3 method for class 'nls':
profile(fitted, which = 1:npar, maxpts = 100, alphamax = 0.01,
        delta.t = cutoff/5, ...)

Arguments

fitted 	 the original fitted model object.

which 	 the original model parameters which should be profiled. This can be a numeric or character vector. By default, all non-linear parameters are profiled.

maxpts 	 maximum number of points to be used for profiling each parameter.

alphamax 	 highest significance level allowed for the profile t-statistics.

delta.t 	 suggested change on the scale of the profile t-statistics. Default value chosen to allow profiling at about 10 parameter values.

... 	 further arguments passed to or from other methods.

Details

The profile t-statistics is defined as the square root of change in sum-of-squares divided by residual standard error with an appropriate sign.

Value

A list with an element for each parameter being profiled. The elements are data-frames with two variables

par.vals 	 a matrix of parameter values for each fitted model.

tau 	 the profile t-statistics.

Author(s)

Of the original version, Douglas M. Bates and Saikat DebRoy

References

Bates, D. M. and Watts, D. G. (1988), Nonlinear Regression Analysis and Its Applications, Wiley (chapter 6).

See Also

nls, profile, plot.profile.nls

Examples

# obtain the fitted object
fm1 <- nls(demand ~ SSasympOrig(Time, A, lrc), data = BOD)
# get the profile for the fitted model: default level is too extreme
pr1 <- profile(fm1, alpha = 0.05)
# profiled values for the two parameters
pr1$A
pr1$lrc
# see also example(plot.profile.nls)

[Package stats version 2.9.2 Index]