sd {stats}	R Documentation

Standard Deviation

Description

This function computes the standard deviation of the values in x. If na.rm is TRUE then missing values are removed before computation proceeds.

Usage


sd(x, na.rm = FALSE)

Arguments

x
a numeric vector or an R object but not a factor coercible to numeric by as.double(x).

na.rm
logical. Should missing values be removed?

Details

Like var this uses denominator n - 1.

The standard deviation of a length-one or zero-length vector is NA.

See Also

var for its square, and mad, the most robust alternative.

Examples


sd(1:2) ^ 2

[Package stats version 3.6.1 Index]