shapiro.test {stats}	R Documentation

Shapiro-Wilk Normality Test

Description

Performs the Shapiro-Wilk test of normality.

Usage

shapiro.test(x)

Arguments

x 	 a numeric vector of data values. Missing values are allowed, but the number of non-missing values must be between 3 and 5000.

Value

A list with class "htest" containing the following components:

statistic 	 the value of the Shapiro-Wilk statistic.

p.value 	 an approximate p-value for the test. This is said in Royston (1995) to be adequate for p.value < 0.1.

method 	 the character string "Shapiro-Wilk normality test".

data.name 	 a character string giving the name(s) of the data.

Source

The algorithm used is a C translation of the Fortran code described in Royston (1995) and found at http://lib.stat.cmu.edu/apstat/R94. The calculation of the p value is exact for n = 3, otherwise approximations are used, separately for 4 <= n <= 11 and n >= 12.

References

Patrick Royston (1982) An extension of Shapiro and Wilk's W test for normality to large samples. Applied Statistics, 31, 115–124.

Patrick Royston (1982) Algorithm AS 181: The W test for Normality. Applied Statistics, 31, 176–180.

Patrick Royston (1995) Remark AS R94: A remark on Algorithm AS 181: The W test for normality. Applied Statistics, 44, 547–551.

See Also

qqnorm for producing a normal quantile-quantile plot.

Examples

shapiro.test(rnorm(100, mean = 5, sd = 3))
shapiro.test(runif(100, min = 2, max = 4))

[Package stats version 2.9.2 Index]