stats-deprecated {stats}	R Documentation

Deprecated Functions in Stats package

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as the next release.

Details

There are currently no deprecated functions in this package.

See Also

Deprecated

[Package stats version 2.9.2 Index]