summary.glm {stats}	R Documentation

Summarizing Generalized Linear Model Fits

Description

These functions are all methods for class glm or summary.glm objects.

Usage

## S3 method for class 'glm':
summary(object, dispersion = NULL, correlation = FALSE,
        symbolic.cor = FALSE, ...)

## S3 method for class 'summary.glm':
print(x, digits = max(3, getOption("digits") - 3),
      symbolic.cor = x$symbolic.cor,
      signif.stars = getOption("show.signif.stars"), ...)

Arguments

object 	 an object of class "glm", usually, a result of a call to glm.

x 	 an object of class "summary.glm", usually, a result of a call to summary.glm.

dispersion 	 the dispersion parameter for the family used. Either a single numerical value or NULL (the default), when it is inferred from object (see ‘Details’).

correlation 	 logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.

digits 	 the number of significant digits to use when printing.

symbolic.cor 	 logical. If TRUE, print the correlations in a symbolic form (see symnum) rather than as numbers.

signif.stars 	 logical. If TRUE, ‘significance stars’ are printed for each coefficient.

... 	 further arguments passed to or from other methods.

Details

print.summary.glm tries to be smart about formatting the coefficients, standard errors, etc. and additionally gives ‘significance stars’ if signif.stars is TRUE. The coefficients component of the result gives the estimated coefficients and their estimated standard errors, together with their ratio. This third column is labelled t ratio if the dispersion is estimated, and z ratio if the dispersion is known (or fixed by the family). A fourth column gives the two-tailed p-value corresponding to the t or z ratio based on a Student t or Normal reference distribution. (It is possible that the dispersion is not known and there are no residual degrees of freedom from which to estimate it. In that case the estimate is NaN.)

Aliased coefficients are omitted in the returned object but restored by the print method.

Correlations are printed to two decimal places (or symbolically): to see the actual correlations print summary(object)$correlation directly.

The dispersion of a GLM is not used in the fitting process, but it is needed to find standard errors. If dispersion is not supplied or NULL, the dispersion is taken as 1 for the binomial and Poisson families, and otherwise estimated by the residual Chisquared statistic (calculated from cases with non-zero weights) divided by the residual degrees of freedom.

summary can be used with Gaussian glm fits to handle the case of a linear regression with known error variance, something not handled by summary.lm.

Value

summary.glm returns an object of class "summary.glm", a list with components

call 	 the component from object.

family 	 the component from object.

deviance 	 the component from object.

contrasts 	 the component from object.

df.residual 	 the component from object.

null.deviance 	 the component from object.

df.null 	 the component from object.

deviance.resid 	 the deviance residuals: see residuals.glm.

coefficients 	 the matrix of coefficients, standard errors, z-values and p-values. Aliased coefficients are omitted.

aliased 	 named logical vector showing if the original coefficients are aliased.

dispersion 	 either the supplied argument or the inferred/estimated dispersion if the latter is NULL.

df 	 a 3-vector of the rank of the model and the number of residual degrees of freedom, plus number of non-aliased coefficients.

cov.unscaled 	 the unscaled (dispersion = 1) estimated covariance matrix of the estimated coefficients.

cov.scaled 	 ditto, scaled by dispersion.

correlation 	 (only if correlation is true.) The estimated correlations of the estimated coefficients.

symbolic.cor 	 (only if correlation is true.) The value of the argument symbolic.cor.

See Also

glm, summary.

Examples

## --- Continuing the Example from  '?glm':

summary(glm.D93)

[Package stats version 2.9.2 Index]