summary.nls {stats}	R Documentation

Summarizing Non-Linear Least-Squares Model Fits

Description

summary method for class "nls".

Usage

## S3 method for class 'nls':
summary(object, correlation = FALSE, symbolic.cor = FALSE, ...)

## S3 method for class 'summary.nls':
print(x, digits = max(3, getOption("digits") - 3),
      symbolic.cor = x$symbolic.cor,
      signif.stars = getOption("show.signif.stars"), ...)

Arguments

object 	 an object of class "nls".

x 	 an object of class "summary.nls", usually the result of a call to summary.nls.

correlation 	 logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.

digits 	 the number of significant digits to use when printing.

symbolic.cor 	 logical. If TRUE, print the correlations in a symbolic form (see symnum) rather than as numbers.

signif.stars 	 logical. If TRUE, ‘significance stars’ are printed for each coefficient.

... 	 further arguments passed to or from other methods.

Details

The distribution theory used to find the distribution of the standard errors and of the residual standard error (for t ratios) is based on linearization and is approximate, maybe very approximate.

print.summary.nls tries to be smart about formatting the coefficients, standard errors, etc. and additionally gives ‘significance stars’ if signif.stars is TRUE.

Correlations are printed to two decimal places (or symbolically): to see the actual correlations print summary(object)$correlation directly.

Value

The function summary.nls computes and returns a list of summary statistics of the fitted model given in object, using the component "formula" from its argument, plus

residuals 	 the weighted residuals, the usual residuals rescaled by the square root of the weights specified in the call to nls.

coefficients 	 a p x 4 matrix with columns for the estimated coefficient, its standard error, t-statistic and corresponding (two-sided) p-value.

sigma 	 the square root of the estimated variance of the random error
sigma^2 = 1/(n-p) Sum(R[i]^2),

where R[i] is the i-th weighted residual.

df 	 degrees of freedom, a 2-vector (p, n-p). (Here and elsewhere n omits observations with zero weights.)

cov.unscaled 	 a p x p matrix of (unscaled) covariances of the parameter estimates.

correlation 	 the correlation matrix corresponding to the above cov.unscaled, if correlation = TRUE is specified and there are a non-zero number of residual degrees of freedom.

symbolic.cor 	 (only if correlation is true.) The value of the argument symbolic.cor.

See Also

The model fitting function nls, summary.

Function coef will extract the matrix of coefficients with standard errors, t-statistics and p-values.

[Package stats version 2.9.2 Index]