summary.princomp {stats}	R Documentation

Summary method for Principal Components Analysis

Description

The summary method for class "princomp".

Usage

## S3 method for class 'princomp':
summary(object, loadings = FALSE, cutoff = 0.1, ...)

## S3 method for class 'summary.princomp':
print(x, digits = 3, loadings = x$print.loadings,
      cutoff = x$cutoff, ...)

Arguments

object 	 an object of class "princomp", as from princomp().

loadings 	 logical. Should loadings be included?

cutoff 	 numeric. Loadings below this cutoff in absolute value are shown as blank in the output.

x 	 an object of class "summary.princomp".

digits 	 the number of significant digits to be used in listing loadings.

... 	 arguments to be passed to or from other methods.

Value

object with additional components cutoff and print.loadings.

See Also

princomp

Examples

summary(pc.cr <- princomp(USArrests, cor=TRUE))
print(summary(princomp(USArrests, cor=TRUE),
              loadings = TRUE, cutoff = 0.2), digits = 2)

[Package stats version 2.9.2 Index]