ts-methods {stats}	R Documentation

Methods for Time Series Objects

Description

Methods for objects of class "ts", typically the result of ts.

Usage

## S3 method for class 'ts':
diff(x, lag = 1, differences = 1, ...)

## S3 method for class 'ts':
na.omit(object, ...)

Arguments

x 	 an object of class "ts" containing the values to be differenced.

lag 	 an integer indicating which lag to use.

differences 	 an integer indicating the order of the difference.

object 	 a univariate or multivariate time series.

... 	 further arguments to be passed to or from methods.

Details

The na.omit method omits initial and final segments with missing values in one or more of the series. ‘Internal’ missing values will lead to failure.

Value

For the na.omit method, a time series without missing values. The class of object will be preserved.

See Also

diff; na.omit, na.fail, na.contiguous.

[Package stats version 2.9.2 Index]