vcov {stats}	R Documentation

Calculate Variance-Covariance Matrix for a Fitted Model Object

Description

Returns the variance-covariance matrix of the main parameters of a fitted model object.

Usage

vcov(object, ...)

Arguments

object 	 a fitted model object.

... 	 additional arguments for method functions. For the glm method this can be used to pass a dispersion parameter.

Details

This is a generic function. Functions with names beginning in vcov. will be methods for this function. Classes with methods for this function include: lm, mlm, glm, nls, lme, gls, coxph and survreg (the last two in package survival).

Value

A matrix of the estimated covariances between the parameter estimates in the linear or non-linear predictor of the model.

[Package stats version 2.1.0 Index]